script to find app-fail list from capture log.
